requirements.txt
Django==5.1.5
drf-yasg==1.21.8
djangorestframework==3.15.2
autoutils==1.0.0
django-constance==4.2.0
python-decouple==3.8
python-telegram-bot==21.10
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
django-autoutils==2.1.1
django-redis==5.4.0
gunicorn==23.0.0